requirements.txt
asgiref==3.7.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-filter==23.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.6
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.11.1
flake8==6.0.0
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pep8-naming==0.13.3
Pillow==9.5.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pytest==7.3.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==2.0.3
